add selected check mark for background
